requirements.txt
transformers
datasets
sentence-transformers
accelerate
bitsandbytes
einops
faiss-cpu
haystack-ai
ragas
tiktoken
requests
tqdm